Adding DpsGetTTLWithReconnect and DpsPutTTLWithReconnect composites
DpsGetTTLComposite - Uses the dpsGetTTL function to read key-value pairs
from DPS (of any type). The purpose of this composite is to provide
extensive error handling, a reconnection policy, and a complete set of
informative metrics. 

DpsPutTTLComposite - Uses the dpsPutTTL function to write key-value
pairs to Redis (of any type). The purpose of this composite is to
provide extensive error handling, a reconnection policy, and a complete
set of informative metrics.