Making Feedback loop handling more robust
I only ever saw the feedback loop cause a locking situation when things
were fused, so originally I simply put the MetricsAggregate in a
partitionIsolation configuration. This seemed to make the problem go
away, but in theory this could happen in separate processes too. By
putting a drop on the input port of the feedback loop queue, we know
that we can't enter a deadlock situation.